Fixed orientation problems and navigation bugs.
Fixed bugs with rotation of the app. Locked orientation of CameraFragment, full sensor orientation for everything else. Fixed bug with opening activity twice when it's working on background mode and open it with widget. Reused function opening the photo in local Gallery. Saved current fragment and load it when return from background mode. Added onInfoWindowClick(...) opening the photo in Gallery. Minor changer to UI.